requirements.txt
appnope==0.1.0
backcall==0.1.0
decorator==4.3.0
dj-database-url==0.5.0
Django==2.1.2
djangorestframework==3.8.2
gunicorn==19.9.0
ipython==7.0.1
ipython-genutils==0.2.0
jedi==0.13.1
mysqlclient==1.3.13
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.5
psycopg2==2.7.5
ptyprocess==0.6.0
Pygments==2.2.0
python-decouple==3.1
pytz==2018.5
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
whitenoise==4.1
